requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
pydantic
pydantic-settings

python-jose[cryptography]
passlib[bcrypt]

python-multipart

python-dotenv

email-validator
bcrypt==4.0.1